RAG Based Chatbot/server/python_engine/requirements.txt
groq
numpy
pymongo
requests
beautifulsoup4
huggingface_hub
python-dotenv
pdfplumber
